requirements.txt
pytest
pytest-playwright
pytest-xdist
python-dotenv
faker